adicionado: Tratamento de exceções de findbyId e delete

+++ src/main/java/com/aprendendojava/springjava/services/UserService.java
@@ -4,10 +4,14 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.aprendendojava.springjava.entities.User;
 import com.aprendendojava.springjava.repositories.UserRepository;
+import com.aprendendojava.springjava.services.exceptions.DatabaseException;
+import com.aprendendojava.springjava.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
@@ -21,7 +25,7 @@ public class UserService {
 	
 	public User findById(Long id) {
 		Optional<User> user = userRepository.findById(id);
-		return user.get();
+		return user.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 	
 	public User insert(User user) {
@@ -29,7 +33,17 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		try {			
+			if(userRepository.existsById(id)) {
+				userRepository.deleteById(id);				
+			}else {
+				throw new ResourceNotFoundException(id);
+			}
+		}catch(EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {			
+	        throw new DatabaseException(e.getMessage());		
+	    }	
 	}
 	
 	public User update(Long id, User user) {
